requirements.txt
openai>=0.27.0
hydra-core
gymnasium
protobuf<=3.20.99
scikit-learn
langchain>=0.0.200
faiss-cpu
sentence_transformers
black
python-dotenv
numpy
pandas
matplotlib
seaborn
pillow
scipy
tokenizers
torch>=2.0.0
transformers>=4.30.0
tqdm
tiktoken
pinecone-client
omegaconf 
langchain-openai